// src/api/generated/models/MenuResponse.ts
export interface MenuResponse {
    /**
     * 
     * @type {number}
     * @memberof MenuResponse
     */
    menuNo?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuResponse
     */
    menuCode?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuResponse
     */
    menuName?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuResponse
     */
    menuType?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuResponse
     */
    menuPath?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuResponse
     */
    apiEndpoint?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuResponse
     */
    iconName?: string;
    /**
     * 
     * @type {number}
     * @memberof MenuResponse
     */
    upperMenuNo?: number;
    /**
     * 
     * @type {number}
     * @memberof MenuResponse
     */
    menuOrder?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuResponse
     */
    menuDescription?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuResponse
     */
    isVisible?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuResponse
     */
    isActive?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {Date}
     * @memberof MenuResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof MenuResponse
     */
    updatedDate?: Date;
    /**
     * 
     * @type {Array<MenuResponse>}
     * @memberof MenuResponse
     */
    children?: Array<MenuResponse>;
}